[WIp] MOVING several things to adapter module proposals_for_votainteligente
