+++ popular_proposal/tests/liking_proposals_tests.py
@@ -14,7 +14,6 @@ class SubscribingToPopularProposal(TestCase):
     def setUp(self):
         super(SubscribingToPopularProposal, self).setUp()
         self.proposal = PopularProposal.objects.create(proposer=self.fiera,
-                                                       area=self.arica,
                                                        data=self.data,
                                                        title=u'This is a title'
                                                        )
