+++ popular_proposal/tests/popular_proposal_model_tests.py
@@ -1,6 +1,5 @@
 # coding=utf-8
 from popular_proposal.tests import ProposingCycleTestCaseBase
-from elections.models import Area
 from django.contrib.auth.models import User
 from popular_proposal.models import ProposalTemporaryData, PopularProposal, ProposalLike
 from popular_proposal.forms.form_texts import TOPIC_CHOICES
@@ -19,7 +18,6 @@ class PopularProposalTestCase(ProposingCycleTestCaseBase):
 
     def test_instantiate_one(self):
         popular_proposal = PopularProposal.objects.create(proposer=self.fiera,
-                                                          area=self.arica,
                                                           data=self.data,
                                                           title=u'This is a title',
                                                           clasification=u'education'
@@ -29,7 +27,6 @@ class PopularProposalTestCase(ProposingCycleTestCaseBase):
         self.assertTrue(popular_proposal.slug)
         self.assertEquals(popular_proposal.title, u'This is a title')
         self.assertIn(popular_proposal, self.fiera.popularproposals.all())
-        self.assertIn(popular_proposal, self.arica.popularproposals.all())
         self.assertIsNone(popular_proposal.temporary)
         self.assertFalse(popular_proposal.background)
         self.assertFalse(popular_proposal.contact_details)
@@ -49,7 +46,6 @@ class PopularProposalTestCase(ProposingCycleTestCaseBase):
 
     def test_popular_proposal_card_as_property(self):
         popular_proposal = PopularProposal.objects.create(proposer=self.fiera,
-                                                          area=self.arica,
                                                           data=self.data,
                                                           title=u'This is a title',
                                                           clasification=u'education'
@@ -62,14 +58,12 @@ class PopularProposalTestCase(ProposingCycleTestCaseBase):
 
     def test_get_classification_text(self):
         popular_proposal = PopularProposal.objects.create(proposer=self.fiera,
-                                                       area=self.arica,
                                                        data=self.data,
                                                        title=u'This is a title',
                                                        clasification=TOPIC_CHOICES[1][0]
                                                        )
         self.assertEquals(popular_proposal.get_classification(), TOPIC_CHOICES[1][1])
         popular_proposal = PopularProposal.objects.create(proposer=self.fiera,
-                                                       area=self.arica,
                                                        data=self.data,
                                                        title=u'This is a title',
                                                        clasification="perrito"
@@ -78,13 +72,11 @@ class PopularProposalTestCase(ProposingCycleTestCaseBase):
 
     def test_reportedproposals_are_not_in_default_manager(self):
         p1 = PopularProposal.objects.create(proposer=self.fiera,
-                                            area=self.arica,
                                             data=self.data,
                                             title=u'This is a title',
                                             clasification=u'education'
                                             )
         p2 = PopularProposal.objects.create(proposer=self.fiera,
-                                            area=self.arica,
                                             data=self.data,
                                             title=u'This is a title',
                                             clasification=u'education',
@@ -101,20 +93,17 @@ class PopularProposalTestCase(ProposingCycleTestCaseBase):
 
     def test_featured_proposals_are_first(self):
         p1 = PopularProposal.objects.create(proposer=self.fiera,
-                                            area=self.arica,
                                             data=self.data,
                                             title=u'This is a title1',
                                             clasification=u'education'
                                             )
         p2 = PopularProposal.objects.create(proposer=self.fiera,
-                                            area=self.arica,
                                             data=self.data,
                                             title=u'This is a title2',
                                             clasification=u'education',
                                             featured=True
                                             )
         p3 = PopularProposal.objects.create(proposer=self.fiera,
-                                            area=self.arica,
                                             data=self.data,
                                             title=u'This is a title3',
                                             clasification=u'education'
@@ -124,20 +113,17 @@ class PopularProposalTestCase(ProposingCycleTestCaseBase):
 
     def test_home_view_brings_featured_first(self):
         p1 = PopularProposal.objects.create(proposer=self.fiera,
-                                            area=self.arica,
                                             data=self.data,
                                             title=u'This is a title1',
                                             clasification=u'education'
                                             )
         p2 = PopularProposal.objects.create(proposer=self.fiera,
-                                            area=self.arica,
                                             data=self.data,
                                             title=u'This is a title2',
                                             clasification=u'education',
                                             featured=True
                                             )
         p3 = PopularProposal.objects.create(proposer=self.fiera,
-                                            area=self.arica,
                                             data=self.data,
                                             title=u'This is a title3',
                                             clasification=u'education',
@@ -153,7 +139,6 @@ class PopularProposalTestCase(ProposingCycleTestCaseBase):
 
     def test_get_ribbon_text(self):
         p1 = PopularProposal.objects.create(proposer=self.fiera,
-                                            area=self.arica,
                                             data=self.data,
                                             title=u'This is a title1',
                                             clasification=u'education'
@@ -166,20 +151,17 @@ class PopularProposalTestCase(ProposingCycleTestCaseBase):
     @override_settings(EXCLUDED_PROPOSALS_APPS=["sites" ,])
     def test_proposals_exclude_certain_types_of_proposals(self):
         p1 = PopularProposal.objects.create(proposer=self.fiera,
-                                            area=self.arica,
                                             data=self.data,
                                             title=u'This is a title1',
                                             clasification=u'education'
                                             )
         p2 = PopularProposal.objects.create(proposer=self.fiera,
-                                            area=self.arica,
                                             data=self.data,
                                             title=u'This is a title2',
                                             clasification=u'education',
                                             featured=True
                                             )
         p3 = PopularProposal.objects.create(proposer=self.fiera,
-                                            area=self.arica,
                                             data=self.data,
                                             title=u'This is a title3',
                                             clasification=u'education'
@@ -190,26 +172,9 @@ class PopularProposalTestCase(ProposingCycleTestCaseBase):
         proposals = PopularProposal.objects.all()
         self.assertNotIn(p3, proposals)
 
-    def test_proposal_ogp(self):
-        site = Site.objects.get_current()
-        popular_proposal = PopularProposal.objects.create(proposer=self.fiera,
-                                                          area=self.arica,
-                                                          data=self.data,
-                                                          title=u'This is a title',
-                                                          clasification=u'education'
-                                                          )
-        self.assertTrue(popular_proposal.ogp_enabled)
-        self.assertTrue(popular_proposal.ogp_title())
-        self.assertEquals('website', popular_proposal.ogp_type())
-        expected_url = "http://%s%s" % (site.domain,
-                                        popular_proposal.get_absolute_url())
-        self.assertEquals(expected_url, popular_proposal.ogp_url())
-        self.assertTrue(popular_proposal.ogp_image())
-
 
     def test_generate_og_image(self):
         popular_proposal = PopularProposal.objects.create(proposer=self.fiera,
-                                                          area=self.arica,
                                                           data=self.data,
                                                           title=u'This is a title',
                                                           clasification=u'education'
@@ -228,7 +193,6 @@ class PopularProposalTestCase(ProposingCycleTestCaseBase):
         popular_proposal = temporary_data.create_proposal(moderator=self.feli)
         self.assertEquals(popular_proposal.proposer, self.fiera)
         popular_proposal = PopularProposal.objects.get(id=popular_proposal.id)
-        self.assertEquals(popular_proposal.area, self.arica)
         self.assertEquals(popular_proposal.join_advocacy_url, u"http://whatsapp.com/somegroup")
         self.assertEquals(popular_proposal.clasification, data['clasification'])
         self.assertEquals(popular_proposal.data, self.data)
@@ -243,32 +207,8 @@ class PopularProposalTestCase(ProposingCycleTestCaseBase):
         self.assertIn(self.fiera.email, the_mail.to)
         self.assertEquals(len(the_mail.to), original_amount_of_mails + 1)
 
-        # context = Context({'area': self.arica,
-        #                    'temporary_data': temporary_data,
-        #                    'moderator': self.feli
-        #                    })
-        # template_body = get_template('mails/popular_proposal_accepted_body.html')
-        # template_subject = get_template('mails/popular_proposal_accepted_subject.html')
-        # expected_content= template_body.render(context)
-        # expected_subject = template_subject.render(context)
-        # self.assertTrue(the_mail.body)
-        # self.assertTrue(the_mail.subject)
-        # self.assertIn(self.data['title'], str(popular_proposal))
-
-    def test_proposal_has_where_was_generated(self):
-        a_comuna = Area.objects.filter(classification='Comuna').first()
-        popular_proposal = PopularProposal.objects.create(proposer=self.fiera,
-                                                          area=self.arica,
-                                                          data=self.data,
-                                                          generated_at=a_comuna,
-                                                          title=u'This is a title',
-                                                          clasification=u'education'
-                                                          )
-        self.assertTrue(popular_proposal)
-
     def test_proposal_has_supporting_organizations(self):
         popular_proposal = PopularProposal.objects.create(proposer=self.fiera,
-                                                          area=self.arica,
                                                           data=self.data,
                                                           title=u'This is a title'
                                                           )
@@ -301,7 +241,6 @@ class PopularProposalTestCase(ProposingCycleTestCaseBase):
 
     def test_render_proposal_detail(self):
         popular_proposal = PopularProposal.objects.create(proposer=self.fiera,
-                                                          area=self.arica,
                                                           data=self.data,
                                                           title=u'This is a title'
                                                           )
@@ -313,7 +252,6 @@ class PopularProposalTestCase(ProposingCycleTestCaseBase):
 
     def test_get_oneliner(self):
         popular_proposal = PopularProposal.objects.create(proposer=self.fiera,
-                                                          area=self.arica,
                                                           data=self.data,
                                                           title=u'This is a title'
                                                           )
